Show the item detail in a modal instead of a dedicated route
Replaced the /items/[itemUuid] page with an ItemDetailDialog that
wraps the same ItemDetailPane two-column view in a full-screen modal.
Clicking an item in the items tab now opens the modal without leaving
the page or remounting the items table.
