--- src/app/human-alignment/tasks/[uuid]/items/[itemUuid]/layout.tsx
@@ -1,13 +1,0 @@
-import { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Item | Calibrate",
-};
-
-export default function LabellingItemLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return children;
-}
