+++ src/components/human-labelling/ItemDetailDialog.tsx
@@ -1,0 +1,311 @@
+"use client";
+
+import { useCallback, useEffect, useMemo, useState } from "react";
+import { apiClient } from "@/lib/api";
+import { type Item } from "@/components/human-labelling/AnnotationJobView";
+import {
+  ItemDetailPane,
+  extractEvaluatorVariables,
+  type EvaluatorRunRow,
+  type HumanAgreementItem,
+  type HumanAnnotation,
+} from "@/components/human-labelling/EvaluatorRunDetailView";
+
+type SummaryAnnotator = { uuid: string; name: string };
+type SummaryAnnotation = {
+  value: boolean | number | null;
+  reasoning?: string | null;
+};
+type SummaryRow = {
+  item_id: string;
+  payload: Record<string, unknown> | null;
+  evaluator_id: string;
+  evaluator_name: string;
+  output_type: "binary" | "rating";
+  evaluator_version_id?: string | null;
+  evaluator_version_number?: number | null;
+  evaluator_value: boolean | number | null;
+  evaluator_reasoning?: string | null;
+  human_agreement: number | null;
+  evaluator_agreement: number | null;
+  annotations: Record<string, SummaryAnnotation | null>;
+};
+type TaskSummaryResponse = {
+  annotators: SummaryAnnotator[];
+  rows: SummaryRow[];
+};
+
+type TaskEvaluatorDef = {
+  uuid: string;
+  description?: string | null;
+  scale_min?: number | boolean | null;
+  scale_max?: number | boolean | null;
+};
+
+export type ItemDetailDialogTask = {
+  uuid: string;
+  name: string;
+  type: "llm" | "stt" | "simulation";
+  evaluators?: TaskEvaluatorDef[];
+};
+
+export function ItemDetailDialog({
+  isOpen,
+  onClose,
+  task,
+  item,
+  accessToken,
+}: {
+  isOpen: boolean;
+  onClose: () => void;
+  task: ItemDetailDialogTask | null;
+  item: Item | null;
+  accessToken: string | null;
+}) {
+  const [summary, setSummary] = useState<TaskSummaryResponse | null>(null);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const taskUuid = task?.uuid;
+  const itemUuid = item?.uuid;
+
+  const fetchSummary = useCallback(async () => {
+    if (!accessToken || !taskUuid || !itemUuid) return;
+    setLoading(true);
+    setError(null);
+    setSummary(null);
+    try {
+      const data = await apiClient<TaskSummaryResponse>(
+        `/annotation-tasks/${taskUuid}/summary?item_id=${encodeURIComponent(itemUuid)}`,
+        accessToken,
+      );
+      setSummary(data);
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : "Failed to load item";
+      setError(msg);
+    } finally {
+      setLoading(false);
+    }
+  }, [accessToken, taskUuid, itemUuid]);
+
+  useEffect(() => {
+    if (isOpen) fetchSummary();
+  }, [isOpen, fetchSummary]);
+
+  // Close on Escape.
+  useEffect(() => {
+    if (!isOpen) return;
+    const handler = (e: KeyboardEvent) => {
+      if (e.key === "Escape") onClose();
+    };
+    window.addEventListener("keydown", handler);
+    return () => window.removeEventListener("keydown", handler);
+  }, [isOpen, onClose]);
+
+  const evaluatorVariables = useMemo(
+    () => (item ? extractEvaluatorVariables(item.payload) : {}),
+    [item],
+  );
+
+  const adapted = useMemo(() => {
+    if (!summary || !task || !item) return null;
+    const taskEvaluatorByUuid = new Map(
+      (task.evaluators ?? []).map((e) => [e.uuid, e]),
+    );
+    const annotatorNameById = new Map(
+      (summary.annotators ?? []).map((a) => [a.uuid, a.name]),
+    );
+
+    const evaluators: {
+      evaluator_id: string;
+      evaluator_version_id?: string;
+      name?: string;
+    }[] = [];
+    const seenEvKey = new Set<string>();
+    const runs: EvaluatorRunRow[] = [];
+    const versionLabels: Record<string, string> = {};
+    const evaluatorNamesById: Record<string, string> = {};
+    const haEvaluators: HumanAgreementItem["evaluators"] = [];
+
+    for (const row of summary.rows) {
+      const evKey = `${row.evaluator_id}-${row.evaluator_version_id ?? ""}`;
+      if (!seenEvKey.has(evKey)) {
+        seenEvKey.add(evKey);
+        evaluators.push({
+          evaluator_id: row.evaluator_id,
+          evaluator_version_id: row.evaluator_version_id ?? undefined,
+          name: row.evaluator_name,
+        });
+      }
+      if (
+        row.evaluator_version_id &&
+        typeof row.evaluator_version_number === "number"
+      ) {
+        versionLabels[row.evaluator_version_id] =
+          `v${row.evaluator_version_number}`;
+      }
+      if (!evaluatorNamesById[row.evaluator_id]) {
+        evaluatorNamesById[row.evaluator_id] = row.evaluator_name;
+      }
+
+      const taskEv = taskEvaluatorByUuid.get(row.evaluator_id);
+      const scaleMin =
+        typeof taskEv?.scale_min === "number" ? taskEv.scale_min : null;
+      const scaleMax =
+        typeof taskEv?.scale_max === "number" ? taskEv.scale_max : null;
+
+      runs.push({
+        uuid: `${row.item_id}:${row.evaluator_id}:${row.evaluator_version_id ?? ""}`,
+        job_id: "",
+        item_id: row.item_id,
+        evaluator_id: row.evaluator_id,
+        evaluator_version_id: row.evaluator_version_id ?? "",
+        value:
+          row.evaluator_value === null && !row.evaluator_reasoning
+            ? null
+            : {
+                value: row.evaluator_value,
+                reasoning: row.evaluator_reasoning ?? null,
+              },
+        status: row.evaluator_value !== null ? "completed" : "pending",
+        created_at: "",
+        completed_at: null,
+        evaluator_version: {
+          uuid: row.evaluator_version_id ?? undefined,
+          version_number: row.evaluator_version_number ?? undefined,
+          scale_min: scaleMin,
+          scale_max: scaleMax,
+        },
+        evaluator: {
+          uuid: row.evaluator_id,
+          name: row.evaluator_name,
+          description: taskEv?.description ?? null,
+          output_type: row.output_type,
+        },
+      });
+
+      const human_annotations: HumanAnnotation[] = [];
+      for (const [annUuid, ann] of Object.entries(row.annotations ?? {})) {
+        if (!ann || ann.value === null || ann.value === undefined) continue;
+        human_annotations.push({
+          annotation_id: `${row.evaluator_id}:${row.evaluator_version_id ?? ""}:${annUuid}`,
+          annotator_id: annUuid,
+          annotator_name: annotatorNameById.get(annUuid) ?? null,
+          job_id: "",
+          value: { value: ann.value, reasoning: ann.reasoning ?? null },
+          reasoning: ann.reasoning ?? null,
+          updated_at: "",
+        });
+      }
+      haEvaluators.push({
+        evaluator_id: row.evaluator_id,
+        agreement: row.evaluator_agreement,
+        pair_count: human_annotations.length,
+        human_annotations,
+        human_agreement: row.human_agreement,
+        evaluator_agreement: row.evaluator_agreement,
+      });
+    }
+
+    const humanAgreementForItem: HumanAgreementItem = {
+      item_id: item.uuid,
+      annotator_count: (summary.annotators ?? []).length,
+      evaluators: haEvaluators,
+    };
+
+    return {
+      evaluators,
+      evaluatorNamesById,
+      runs,
+      versionLabels,
+      humanAgreementForItem,
+    };
+  }, [summary, task, item]);
+
+  if (!isOpen) return null;
+
+  return (
+    <div
+      className="fixed inset-0 z-50 flex items-center justify-center p-0 md:p-4 bg-black/50 backdrop-blur-sm"
+      onClick={onClose}
+    >
+      <div
+        onClick={(e) => e.stopPropagation()}
+        className="bg-background rounded-none md:rounded-xl w-full max-w-[92rem] h-full md:h-[92vh] flex flex-col shadow-2xl overflow-hidden"
+      >
+        <div className="flex items-center justify-between px-4 md:px-6 py-3 md:py-4 border-b border-border">
+          <div className="min-w-0">
+            <h2 className="text-base md:text-lg font-semibold text-foreground truncate">
+              {task?.name ?? "Item"}
+            </h2>
+            <p className="text-xs text-muted-foreground">Item view</p>
+          </div>
+          <button
+            onClick={onClose}
+            aria-label="Close"
+            className="flex items-center justify-center w-8 h-8 rounded-md hover:bg-muted transition-colors cursor-pointer"
+          >
+            <svg
+              className="w-5 h-5"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+              strokeWidth={2}
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                d="M6 18L18 6M6 6l12 12"
+              />
+            </svg>
+          </button>
+        </div>
+
+        <div className="flex-1 flex flex-col min-h-0 overflow-hidden">
+          {error ? (
+            <div className="m-4 rounded-md border border-border bg-muted/20 p-4 text-sm text-red-500">
+              {error}
+            </div>
+          ) : loading || !task || !item || !adapted ? (
+            <div className="flex-1 flex items-center justify-center gap-2 text-sm text-muted-foreground">
+              <svg
+                className="w-4 h-4 animate-spin"
+                fill="none"
+                viewBox="0 0 24 24"
+              >
+                <circle
+                  className="opacity-25"
+                  cx="12"
+                  cy="12"
+                  r="10"
+                  stroke="currentColor"
+                  strokeWidth="4"
+                />
+                <path
+                  className="opacity-75"
+                  fill="currentColor"
+                  d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+                />
+              </svg>
+              Loading item
+            </div>
+          ) : (
+            <ItemDetailPane
+              item={item}
+              taskType={task.type}
+              evaluators={adapted.evaluators}
+              evaluatorNamesById={adapted.evaluatorNamesById}
+              runs={adapted.runs}
+              versionLabels={adapted.versionLabels}
+              jobStatus="completed"
+              humanAgreementForItem={adapted.humanAgreementForItem}
+              evaluatorVariablesByEvaluatorId={evaluatorVariables}
+              filterDisagreements={false}
+              linkEvaluators
+            />
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
